feat: Finalize initial version with updated data types and documentation.
