+++ .gitignore
@@ -2,4 +2,3 @@
 /lib
 
 .DS_Store
-tsconfig.tsbuildinfo
